% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Summay__Capability_FUNs.R
\name{capability.summary}
\alias{capability.summary}
\title{Calculate Summary of Quality Performance Parameters}
\usage{
capability.summary(LSL, USL, QC.Center, QC.Sigma, s.Sigma, digits = 2)
}
\arguments{
\item{LSL}{number, customer's lower specification limit.}

\item{USL}{number, customer's upper specification limit.}

\item{QC.Center}{number, the mean or median value determined from an XmR plot or a Studentized (e.g., xBar) analysis.}

\item{s.Sigma}{number, the sigma value determined from overall standard deviation (i.e., sd()).}

\item{digits}{integer, how many digits to report.}
}
\value{
data frame , listing of metric lebel and value
}
\description{
Calculate Summary of Quality Performance Parameters
}
