% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUN_Capability.R
\name{Cp}
\alias{Cp}
\title{Calculate Cp}
\usage{
Cp(LSL, USL, QC.Sigma)
}
\arguments{
\item{LSL}{number, customer's lower specification limit.}

\item{USL}{number, customer's upper specification limit.}

\item{QC.Sigma}{number, the sigma value determined from an XmR plot or a Studentized (e.g., xBar) analysis.}
}
\value{
numeric, Cp value (unitless)
}
\description{
function to calculate Cp - "The elbowroom or margin your process"
}
