% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUN_Capability.R
\name{Cpk}
\alias{Cpk}
\title{Calculate Cpk}
\usage{
Cpk(LSL, USL, QC.Center, QC.Sigma)
}
\arguments{
\item{LSL}{number, customer's lower specification limit.}

\item{USL}{number, customer's upper specification limit.}

\item{QC.Center}{number, the mean or median value determined from an XmR plot or a Studentized (e.g., xBar) analysis.}

\item{QC.Sigma}{number, the sigma value determined from an XmR plot or a Studentized (e.g., xBar) analysis.}
}
\value{
numeric, Cpk value (unitless)
}
\description{
function to calculate Cpk - "measure of process centering"
}
