% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{mR_points}
\alias{mR_points}
\title{One Point Moving Range of Vector}
\usage{
mR_points(y)
}
\arguments{
\item{y}{: vector of values}
}
\value{
Vector of one-point moving range.
}
\description{
Calculates a one-point moving range vector given an input vector of values.
 Output often used to produce mR-chart.
}
\examples{
y <- seq(-5:5)
mR_points(y)
}
