% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{npBar_UCL}
\alias{npBar_UCL}
\title{Upper Control Limit: Binomial Data (np-chart)}
\usage{
npBar_UCL(y, n, ...)
}
\arguments{
\item{y}{Vector of binomial count data (not proportions). Each observation
having the same-area of opportunity.}

\item{n}{A number representing the area of opportunity.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A number; 3-sigma upper control limit (UCL)
}
\description{
Calculates upper control limit (UCL) for
binomial count data acquired over the same-sized area of
opportunity.
}
\examples{
set.seed(5555)
p <- rbinom(n = 100, size = 30, prob = .2)
npBar_UCL(y = p, n = 30)

}
