% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{pBar}
\alias{pBar}
\title{Mean Propotion: Binomial Data (p-chart)}
\usage{
pBar(y, n, ...)
}
\arguments{
\item{y}{Vector of binomial proportion data (not counts). Observations
may have a different area of opportinity, n.}

\item{n}{A vector representing the area of opportunity.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector of mean proportion, length equal to length of
parameter y.
}
\description{
Calculates overall mean porportion
for binomial proportion data aquired over a variable area of
opportunity.
}
\examples{
set.seed(5555)
p <- rbinom(n = 100, size = 30, prob = .2)
n <- rpois(100, 100)
pBar(y = p/n, n = n)

}
