% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{pBar_UCL}
\alias{pBar_UCL}
\title{Upper Control Limit: Binomial Data (p-chart)}
\usage{
pBar_UCL(y, n, ...)
}
\arguments{
\item{y}{Vector of binomial proportion data (not counts). Observations
may have a different area of opportunity, n.}

\item{n}{A vector representing the area of opportunity.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector; point-wise 3-sigma upper control limit (UCL)
}
\description{
Calculates point-wise upper control limit (UCL)
for binomial proportion data acquired over a variable area of
opportunity.
}
\examples{
set.seed(5555)
p <- rbinom(n = 100, size = 30, prob = .2)
n <- rpois(100, 100)
pBar_UCL(y = p/n, n = n)

}
