% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUN_Capability.R
\name{Pp}
\alias{Pp}
\title{Calculate Pp}
\usage{
Pp(LSL, USL, s.Sigma)
}
\arguments{
\item{LSL}{number, customer's lower specification limit.}

\item{USL}{number, customer's upper specification limit.}

\item{s.Sigma}{number, the sigma value determined from overall standard deviation (i.e., sd()).}
}
\value{
numeric, Pp value (unitless)
}
\description{
function to calculate Pp - "The elbowroom or margin your process" uses overall sigma value not the QC chart sigma values.
}
