% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUN_Capability.R
\name{process_tolerance}
\alias{process_tolerance}
\title{Calculate QC Process Tolerance}
\usage{
process_tolerance(LSL, USL, QC.Sigma)
}
\arguments{
\item{LSL}{number, customer's lower specification limit.}

\item{USL}{number, customer's upper specification limit.}

\item{QC.Sigma}{number, the sigma value determined from an XmR plot or a Studentized (e.g., xBar) analysis.}
}
\value{
numeric, standardized process tolerance value in sigma units
}
\description{
function to calculate a standardized process tolerance with sigma unit
}
