% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Summay__Capability_FUNs.R
\name{QC_Capability}
\alias{QC_Capability}
\title{Calculate Summary of Quality Performance Parameters}
\usage{
QC_Capability(data = NULL, value = NULL, grouping = NULL,
  formula = NULL, method = "xBar.rBar", na.rm = FALSE, LSL = NULL,
  USL = NULL, digits = 2)
}
\arguments{
\item{data}{vector or dataframe, as indicated below for each chart type
\itemize{
\item \bold{Individuals (XmR)}: vector of values;
\item \bold{Studentized}: dataframe
}}

\item{value}{string, \bold{Studentized Charts}, name of numeric vector in dataframe with values of interest.}

\item{grouping}{string, \bold{Studentized Charts}, name of single factor/variable to split
the dataframe "values" by}

\item{formula}{\bold{Studentized Charts}: a formula,
such as y ~ x1 + x2, where the y variable is
numeric data to be split into groups according to the grouping x
factors/variables}

\item{method}{string, calling one of the following methods:
\itemize{
\item \bold{Individuals Charts}: XmR,
\item \bold{Studentized Charts}: xBar.rBar, xBar.rMedian, xBar.sBar, xMedian.rBar
}}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}

\item{digits}{integer, how many digits to report.}
}
\value{
data frame , listing of metric lebel and value
}
\description{
Calculate Summary of Quality Performance Parameters
}
