% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{QCrange}
\alias{QCrange}
\title{Range: Max Min Difference}
\usage{
QCrange(y)
}
\arguments{
\item{y}{: vector of values}
}
\value{
a number.
}
\description{
Given a set of numbers, function calculates the difference between the maximum and minimum value.
}
\examples{
y <- seq(-5:5)
QCrange(y)
}
