% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{uBar}
\alias{uBar}
\title{Mean Rate: Count Data (u-chart)}
\usage{
uBar(y, n, ...)
}
\arguments{
\item{y}{Vector of counts per unit opportunity (rate). Observations
may have a different area of opportunity, n.}

\item{n}{A vector representing the area of opportunity.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector of mean rate, length equal to length of
parameter y.
}
\description{
Calculates overall mean rate
for count data acquired over a variable area of
opportunity.
}
\examples{
set.seed(5555)
counts <- rpois(100, 25)
n <- rpois(100, 15)
uBar(y = counts / n, n = n)

}
