% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_QC_FUNs.R
\name{xBar_one_UCL}
\alias{xBar_one_UCL}
\title{xBar_One Upper Control Limit (UCL)}
\usage{
xBar_one_UCL(y, na.rm = FALSE, ...)
}
\arguments{
\item{y}{Vector of values}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A number; xBar_One Upper Control Limit (UCL)
}
\description{
Calculates the xBar_One UCL used when constructing a xBar-One chart.
}
\examples{
set.seed(5555)
values <- rnorm(n = 100, mean = 25, sd = 1)
xBar_one_UCL(values)
}
